tightened up validation for event loading
